requirements.txt
# torch==1.7.1
torchvision==0.8.2
matplotlib==3.3.2
scikit-learn==0.23.2
numpy==1.19.2
robustness==1.2.1.post2
tqdm==4.56.0
Pillow==8.1.0
scipy==1.5.2
transformers==4.2.2
umap-learn==0.5.1